Fix: delete multiple entities (#377)
* It is now possible to delete an entity even if it is referenced somewhere else

* Update server/src/main/java/de/uftos/repositories/database/StudentGroupRepository.java

* It is now possibile to delete multiple entities at a time without getting an error (Epecially students and subjects)

* Added a check to not delete a grade if it's still associated with a student group
Now removing student groups from old grades on update

* Update server/src/main/java/de/uftos/services/GradeService.java


* Added delete all methods for tags, teachers, grades and rooms

* Checkstyle

---------

+++ server/src/main/java/de/uftos/services/GradeService.java
@@ -9,7 +9,9 @@ import de.uftos.entities.Lesson;
 import de.uftos.entities.StudentGroup;
 import de.uftos.repositories.database.GradeRepository;
 import de.uftos.repositories.database.ServerRepository;
+import de.uftos.repositories.database.StudentGroupRepository;
 import de.uftos.utils.SpecificationBuilder;
+import java.util.Arrays;
 import java.util.Collection;
 import java.util.List;
 import java.util.Objects;
@@ -30,16 +32,20 @@ import org.springframework.web.server.ResponseStatusException;
 public class GradeService {
   private final GradeRepository repository;
   private final ServerRepository serverRepository;
+  private final StudentGroupRepository studentGroupRepository;
 
   /**
    * Creates a grade service.
    *
-   * @param repository the repository for accessing the grade table.
+   * @param repository             the repository for accessing the grade table.
+   * @param studentGroupRepository the repository for accessing the student group table.
    */
   @Autowired
-  public GradeService(GradeRepository repository, ServerRepository serverRepository) {
+  public GradeService(GradeRepository repository, ServerRepository serverRepository,
+                      StudentGroupRepository studentGroupRepository) {
     this.repository = repository;
     this.serverRepository = serverRepository;
+    this.studentGroupRepository = studentGroupRepository;
   }
 
   /**
@@ -144,10 +150,38 @@ public class GradeService {
     Optional<Grade> grade = this.repository.findById(id);
     if (grade.isEmpty()) {
       throw new ResponseStatusException(HttpStatus.BAD_REQUEST,
-          "Could not find a grade with this id");
+          "Could not find a grade with this id!");
+    }
+    List<StudentGroup> studentGroups = this.studentGroupRepository.findByGrades(grade.get());
+    if (!studentGroups.isEmpty()) {
+      throw new ResponseStatusException(HttpStatus.BAD_REQUEST,
+          "This grade is still associated with a student group.");
     }
 
     this.repository.delete(grade.get());
+  }
+
+  /**
+   * Deletes the grade with the given ID.
+   *
+   * @param ids the IDs of the grades which are to be deleted.
+   * @throws ResponseStatusException is thrown if no grade exists with the given ID.
+   */
+  public void deleteGrades(String[] ids) {
+    List<String> gradesIds = Arrays.asList(ids);
+    List<Grade> grades = this.repository.findAllById(gradesIds);
+    if (grades.size() != gradesIds.size()) {
+      throw new ResponseStatusException(HttpStatus.BAD_REQUEST,
+          "Could not found a grade with this id");
+    }
+
+    List<StudentGroup> studentGroups = this.studentGroupRepository.findAllByGrades(grades);
+    if (!studentGroups.isEmpty()) {
+      throw new ResponseStatusException(HttpStatus.BAD_REQUEST,
+          "This grade is still associated with a student group.");
+    }
+
+    this.repository.deleteAll(grades);
   }
 
   private GradeResponseDto mapResponseDto(Grade grade) {
